Extract model tool execution from JSON runner
Model-callable tools are now a first-class path, so their prompt contracts and tool-call loop should not keep expanding llm_runner.

This patch:

- Moves model-tool prompt guidance, contract clipping, tool-call execution, and configuration errors into model_tool_runner.

- Keeps llm_runner focused on chat model setup, retry behavior, structured output, and JSON parsing.

- Replaces stale llm_runner unit coverage with focused model_tool_runner tests plus one integration test.

- Preserves the existing index_search behavior and provider bind_tools fail-fast semantics.

Constraint: User requested no mypy runs

Rejected: Keep tool-call internals in llm_runner | future MCP/VTX/tree-sitter tools would keep growing the generic JSON runner

Confidence: high

Scope-risk: narrow

Directive: Add future model-tool execution behavior in model_tool_runner, not llm_runner

Tested: uv run ruff check src/metis tests

Tested: uv run pytest -q tests/test_model_tool_runner.py tests/test_llm_runner.py tests/test_engine_core.py tests/test_tool_registry.py tests/test_engine_review.py tests/test_graphs.py tests/test_engine_tools_selection.py tests/test_cli_entry.py

Tested: git diff --check

Not-tested: End-to-end live model call with a provider-backed index_search
