+++ src/metis/engine/llm_runner.py
@@ -5,17 +5,16 @@ from dataclasses import dataclass
 from typing import Any, Callable
 
 from langchain_core.messages import SystemMessage
-from langchain_core.messages import ToolMessage
 from langchain_core.output_parsers import StrOutputParser
 from langchain_core.prompts import ChatPromptTemplate
 
 from metis.chat_model_options import merge_chat_model_kwargs
-
-_MAX_TOOL_CONTRACT_CHARS = 6000
-
-
-class ModelToolConfigurationError(ValueError):
-    pass
+from metis.engine.model_tool_runner import (
+    ModelToolConfigurationError,
+    invoke_model_with_tools,
+    model_tool_system_prompt,
+    require_max_tool_rounds,
+)
 
 
 @dataclass(frozen=True, slots=True)
@@ -47,7 +46,7 @@ class JsonPromptRunner:
     def invoke(self, request: JsonPromptRequest):
         last_failure = "unknown failure"
         max_tool_rounds = (
-            _require_max_tool_rounds(request.max_tool_rounds)
+            require_max_tool_rounds(request.max_tool_rounds)
             if request.model_tools
             else None
         )
@@ -69,7 +68,7 @@ class JsonPromptRunner:
                 if request.temperature is not None:
                     params["temperature"] = request.temperature
                 chat = self._llm_provider.get_chat_model(**params)
-                system_prompt = _system_prompt_with_tool_guidance(
+                system_prompt = model_tool_system_prompt(
                     request.system_prompt,
                     request.model_tools,
                 )
@@ -100,7 +99,7 @@ class JsonPromptRunner:
                         last_failure = f"structured validation failed: {exc}"
                 if parsed is None:
                     if request.model_tools:
-                        response_text = _invoke_with_model_tools(
+                        response_text = invoke_model_with_tools(
                             chat,
                             prompt,
                             request.variables,
@@ -179,126 +178,3 @@ def invoke_langchain_json_prompt_with_retry(
             max_tool_rounds=max_tool_rounds,
         )
     )
-
-
-def _system_prompt_with_tool_guidance(
-    system_prompt: str, tools: tuple[Any, ...]
-) -> str:
-    if not tools:
-        return system_prompt
-    lines = [
-        system_prompt.rstrip(),
-        "",
-        "AVAILABLE MODEL TOOLS",
-        "Use these tools only when they can provide missing project context. "
-        "After any tool calls, return the final response in the requested format.",
-    ]
-    for tool in tools:
-        name = getattr(tool, "name", "")
-        description = getattr(tool, "description", "")
-        if name and description:
-            lines.append(f"- {name}: {description}")
-        elif name:
-            lines.append(f"- {name}")
-    contract_sections = _tool_contract_sections(tools)
-    if contract_sections:
-        lines.extend(["", "MODEL TOOL CONTRACTS", *contract_sections])
-    return "\n".join(lines).strip()
-
-
-def _tool_contract_sections(tools: tuple[Any, ...]) -> list[str]:
-    sections = []
-    for tool in tools:
-        metadata = getattr(tool, "metadata", None) or {}
-        if not isinstance(metadata, dict):
-            continue
-        contract = _clip_tool_contract(str(metadata.get("metis_contract") or ""))
-        if not contract:
-            continue
-        name = getattr(tool, "name", "tool")
-        sections.append(f"[{name}]\n{contract}")
-    return sections
-
-
-def _clip_tool_contract(contract: str) -> str:
-    text = contract.strip()
-    if len(text) <= _MAX_TOOL_CONTRACT_CHARS:
-        return text
-    return text[:_MAX_TOOL_CONTRACT_CHARS].rstrip() + "\n[contract truncated]"
-
-
-def _require_max_tool_rounds(value: int | None) -> int:
-    if value is None:
-        raise ModelToolConfigurationError(
-            "max_tool_rounds must be configured when model_tools are used"
-        )
-    try:
-        max_tool_rounds = int(value)
-    except (TypeError, ValueError) as exc:
-        raise ModelToolConfigurationError(
-            "max_tool_rounds must be a positive integer"
-        ) from exc
-    if max_tool_rounds <= 0:
-        raise ModelToolConfigurationError("max_tool_rounds must be a positive integer")
-    return max_tool_rounds
-
-
-def _invoke_with_model_tools(
-    chat,
-    prompt: ChatPromptTemplate,
-    variables: dict[str, Any],
-    tools: tuple[Any, ...],
-    *,
-    max_tool_rounds: int,
-) -> str:
-    bind_tools = getattr(chat, "bind_tools", None)
-    if not callable(bind_tools):
-        raise ModelToolConfigurationError(
-            "model_tools require a LangChain chat model with bind_tools support"
-        )
-
-    tool_chat = bind_tools(list(tools))
-    tool_by_name = {getattr(tool, "name", ""): tool for tool in tools}
-    messages = prompt.invoke(variables).to_messages()
-    last_response = None
-    for _ in range(max_tool_rounds):
-        last_response = tool_chat.invoke(messages)
-        tool_calls = list(getattr(last_response, "tool_calls", None) or [])
-        if not tool_calls:
-            return _message_content_text(last_response)
-        messages.append(last_response)
-        for index, tool_call in enumerate(tool_calls):
-            name = str(tool_call.get("name") or "")
-            args = tool_call.get("args") or {}
-            tool_call_id = str(tool_call.get("id") or f"{name}-{index}")
-            status = "success"
-            try:
-                tool = tool_by_name[name]
-                content = tool.invoke(args)
-            except Exception as exc:
-                status = "error"
-                content = f"Tool {name!r} failed: {exc}"
-            messages.append(
-                ToolMessage(
-                    content=str(content),
-                    name=name,
-                    tool_call_id=tool_call_id,
-                    status=status,
-                )
-            )
-    return _message_content_text(last_response)
-
-
-def _message_content_text(message: Any) -> str:
-    content = getattr(message, "content", message)
-    if isinstance(content, str):
-        return content
-    if isinstance(content, list):
-        parts = []
-        for item in content:
-            if isinstance(item, str):
-                parts.append(item)
-            elif isinstance(item, dict) and item.get("text") is not None:
-                parts.append(str(item["text"]))
-        return "\n".join(parts)
-    return str(content or "")
